+++ types/date/toEpochMilliseconds.d.ts
@@ -1,0 +1,1 @@
+export { toEpochMilliseconds, toEpochMilliseconds as default } from "../date.js";